+++ Tests/Editor/CrashFixesPolicyTests.cs
@@ -1,0 +1,15 @@
+using NUnit.Framework;
+using ZeyWinAds.Editor.QATests;
+
+namespace ZeyWinAds.Tests.Editor
+{
+    public class CrashFixesPolicyTests
+    {
+        [Test]
+        public void KnownCrashFixes_ArePresent()
+        {
+            var error = CrashFixesPolicy.ValidateAll();
+            Assert.IsNull(error, error);
+        }
+    }
+}